requirements.txt
Flask==3.0.3
gevent==23.9.1
langchain_openai==0.2.2
openai==1.51.2
pandas==2.0.3
PyYAML==6.0.2
Requests==2.32.3
scipy==1.9.3
termcolor==2.5.0
